Add 'About' app and update help link behavior in MenuBar
This commit introduces a new 'About' app to the Ferropoly editor, which includes previously missing sections such as 'Hilfe & Infos', 'Sponsoring', and 'Über diese Software'. Additionally, a modification was made in 'MenuBar.vue' to change the behavior of the help link, removing the target="_blank" attribute to open the link in the same window. The repository URL in 'package.json' has also been updated.